backend/requirements.txt
# === Core Framework ===
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
python-multipart==0.0.20

# === Database ===
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pgvector==0.3.6

# === Redis ===
redis==5.2.1
arq==0.26.1

# === Auth ===
firebase-admin==6.7.0
pyjwt==2.10.1

# === File Storage ===
minio==7.2.15
google-cloud-storage==2.19.0

# === AI / ML ===
google-cloud-aiplatform==1.78.0
google-generativeai==0.8.5

# === File Processing ===
PyPDF2==3.0.1
python-docx==1.1.2
pydub==0.25.1

# === Export ===
weasyprint==63.1

# === Utilities ===
httpx==0.28.1
pydantic[email]==2.11.1
pydantic-settings==2.8.1

# === Testing ===
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
